+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/search/BookmarkPlaceHolder.kt
@@ -1,0 +1,46 @@
+package com.wafflestudio.snutt2.views.logged_in.home.search
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.wafflestudio.snutt2.R
+import com.wafflestudio.snutt2.ui.SNUTTColors
+import com.wafflestudio.snutt2.ui.SNUTTTypography
+
+@Composable
+fun BookmarkPlaceHolder() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(5.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Spacer(modifier = Modifier.weight(1f))
+        Text(
+            text = stringResource(R.string.bookmark_page_placeholder_1),
+            style = SNUTTTypography.subtitle1.copy(
+                fontSize = 18.sp,
+                color = SNUTTColors.White700,
+                fontWeight = FontWeight.Bold,
+            ),
+        )
+        Text(
+            text = stringResource(R.string.bookmark_page_placeholder_2),
+            style = SNUTTTypography.subtitle1.copy(fontSize = 18.sp, color = SNUTTColors.White700),
+        )
+        Text(
+            text = stringResource(R.string.bookmark_page_placeholder_3),
+            style = SNUTTTypography.subtitle1.copy(fontSize = 18.sp, color = SNUTTColors.White700),
+        )
+        Spacer(modifier = Modifier.weight(1f))
+    }
+}
