+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/search/SearchPageMode.kt
@@ -1,0 +1,12 @@
+package com.wafflestudio.snutt2.views.logged_in.home.search
+
+enum class SearchPageMode {
+    Search, Bookmark;
+
+    fun toggled(): SearchPageMode {
+        return when (this) {
+            Search -> Bookmark
+            Bookmark -> Search
+        }
+    }
+}
