+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/search/SearchPlaceHolder.kt
@@ -1,0 +1,89 @@
+package com.wafflestudio.snutt2.views.logged_in.home.search
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.wafflestudio.snutt2.R
+import com.wafflestudio.snutt2.components.compose.BigSearchIcon
+import com.wafflestudio.snutt2.components.compose.clicks
+import com.wafflestudio.snutt2.ui.SNUTTColors
+import com.wafflestudio.snutt2.ui.SNUTTTypography
+
+@Composable
+fun SearchPlaceHolder(onClickSearchIcon: () -> Unit) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(bottom = 40.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+    ) {
+        BigSearchIcon(
+            modifier = Modifier
+                .width(78.dp)
+                .height(76.dp)
+                .padding(10.dp)
+                .clicks {
+                    onClickSearchIcon.invoke()
+                },
+        )
+        Spacer(modifier = Modifier.height(25.dp))
+        Text(
+            text = stringResource(R.string.search_result_placeholder_1),
+            style = SNUTTTypography.h1.copy(fontSize = 25.sp, color = SNUTTColors.White700),
+        )
+        Spacer(modifier = Modifier.height(35.dp))
+        Text(
+            text = stringResource(R.string.search_result_placeholder_2),
+            style = SNUTTTypography.subtitle1.copy(fontSize = 18.sp, color = SNUTTColors.White700),
+        )
+        Spacer(modifier = Modifier.height(5.dp))
+        Text(
+            text = stringResource(R.string.search_result_placeholder_3),
+            style = SNUTTTypography.subtitle2.copy(color = SNUTTColors.White500),
+        )
+        Spacer(modifier = Modifier.height(25.dp))
+        Text(
+            text = stringResource(R.string.search_result_placeholder_4),
+            style = SNUTTTypography.subtitle1.copy(fontSize = 18.sp, color = SNUTTColors.White700),
+        )
+        Spacer(modifier = Modifier.height(5.dp))
+        Text(
+            text = stringResource(R.string.search_result_placeholder_5),
+            style = SNUTTTypography.subtitle2.copy(color = SNUTTColors.White500),
+        )
+    }
+}
+
+@Composable
+fun SearchEmptyPlaceholder() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(bottom = 40.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+    ) {
+        BigSearchIcon(
+            modifier = Modifier
+                .width(58.dp)
+                .height(56.dp),
+        )
+        Spacer(modifier = Modifier.height(35.dp))
+        Text(
+            text = stringResource(R.string.search_result_empty),
+            style = SNUTTTypography.subtitle1.copy(color = SNUTTColors.White700, fontSize = 18.sp),
+        )
+    }
+}
